requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.2
Pillow==5.3.0
PyMuPDF==1.14.13
Werkzeug==0.15.2
